# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VENUS POS API",
        "version": "4.0.0"
    },
    "paths": {
...
        "/api/reports/period": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Report by Period",
                "description": "Previous Route: /enablerAPI/reportsCtrl/getPeriodReport",
                "operationId": "App\\Http\\Controllers\\Api\\PeriodReportController::periodReport",
                "parameters": [
                    {
                        "name": "posID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "periodID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "number",
                                            "example": "int|string|object|array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "Not master pos | Period details not foun | Failed to close"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/period-on-close": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Report Period on close",
                "description": "Previous Route: /enablerAPI/reportsCtrl/getPeriodReportOnClose",
                "operationId": "App\\Http\\Controllers\\Api\\PeriodReportController::periodOnClose",
                "parameters": [
                    {
                        "name": "posID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "number",
                                            "example": "int|string|object|array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "No active period found"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/cdraw-history-by-cdraw-period": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Report Period on close",
                "description": "Previous Route: /enablerAPI/reportsCtrl/getCDrawHistByCDPeriod",
                "operationId": "App\\Http\\Controllers\\Api\\PeriodReportController::cdrawHistoryByCdrawPeriod",
                "parameters": [
                    {
                        "name": "cdrawPeriodID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "number",
                                            "example": "int|string|object|array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "No active period found"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/charge-account-transaction-post-pay": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Charge Account Transaction Report Post Pay",
                "description": "Previous Route: /enablerAPI/reportsCtrl/getChargeAccTransReportPostPay",
                "operationId": "App\\Http\\Controllers\\Api\\PeriodReportController::getChargeAccountTransactionReportPostPay",
                "parameters": [
                    {
                        "name": "periodID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "number",
                                            "example": "int|string|object|array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "statusDescription": {
                                            "type": "string",
                                            "example": "No active period found"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}